feat(data): fix dataset init

+++ src/data/dataset.py
@@ -5,8 +5,8 @@ from torch.utils.data import Dataset
 class ImageCatpionDataset(Dataset):
     def __init__(self, images, captions, targets, start=0, end=1):
         image_num = len(images)
-        start_index = image_num * start
-        end_index = image_num * end
+        start_index = int(image_num * start)
+        end_index = int(image_num * end)
         self.images = images[start_index:end_index]
         self.captions = captions[start_index*5:end_index*5]
         self.targets = targets[start_index*5:end_index*5]
